Bugfix 3.1: changing read's int into size_t

+++ shramko.maksim/P3/inputMtrx.cpp
@@ -1,6 +1,6 @@
 #include "inputMtrx.hpp"
 
-std::istream & shramko::inputMtrx(std::istream & in, int* mtrx, size_t cnt_row, size_t cnt_col, int& read)
+std::istream & shramko::inputMtrx(std::istream & in, int* mtrx, size_t cnt_row, size_t cnt_col, size_t& read)
 {
   for (size_t i = 0; i < (cnt_row * cnt_col); ++i)
   {
